Change to install only one package at a time
Applies to #1, #2

Since the packages are already being installed asynchronously, the
theory is that what is causing the user's Atom to stop is that too many
packages are being installed at once, bogging it down. This change
makes it so that a list of packages is kept and each one is shifted off
the list and installed in turn until the list of packages is empty.